pyproject.toml
[tool.poetry]
name = "ma-min-app"
version = "0.1.0"
description = "Minimal Master Architect app"
authors = ["Ravindra Bajpai <ravindra.bajpai@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.115"
uvicorn = {extras=["standard"], version="^0.30"}
streamlit = "^1.39"
chromadb = "^0.5"
sentence-transformers = "^3.0"
httpx = "^0.27"
pydantic = "^2.7"
python-dotenv = "^1.0"
openai = "^1.50"
requests = "^2.32"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
ruff = "^0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

